remove relationship type field; add indexes to migrations; add metadatum value model

+++ lib/tasks/db_reload.rake
@@ -48,7 +48,6 @@ namespace :files do
   task :clear => :environment do
     dir = "public/" + V1::AssetUploader::BASE_DIR
     puts "Deleting files directory #{dir}"
-    FileUtils.rm_r dir 
-
+    FileUtils.rm_r dir if File.exists?(dir) 
   end
 end
